% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.search.function.R
\name{distribution.search.function}
\alias{distribution.search.function}
\title{A distribution search function}
\usage{
distribution.search.function(x, k, incv, look, upper.prev, lower.prev,
  n.grid, density.prev, target)
}
\arguments{
\item{x}{integrating over x}

\item{k}{the number of treaments in this trial}

\item{incv}{Incv is the number of increments in the integration}

\item{look}{the current interim analysis}

\item{upper.prev}{previous upper bound}

\item{lower.prev}{previous lower bound}

\item{n.grid}{number of points to integrate over}

\item{density.prev}{previous density of test statistic}

\item{target}{target value in the spending function}
}
\description{
This function isn't generally used, it is used to get the distribution of the test statistic at each look
}
