% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.boundaries.R
\name{get.boundaries}
\alias{get.boundaries}
\title{Function to create the boudnaries that satisfy the spending function}
\usage{
get.boundaries(n.looks = 2, v, k = c(2, 1), alpha.star.u = c(0,
  0.025), alpha.star.l = c(0, 0.975))
}
\arguments{
\item{n.looks}{The number of interim analyses occuring in the trial}

\item{v}{Ratio used in the calculation of boundaries}

\item{k}{Number of treatments used in the trial}

\item{alpha.star.u}{the upper limit alpha on boundaries for the spending function}

\item{alpha.star.l}{the lower limit alpha on boundaries for the spending function}
}
\description{
Function to create the boudnaries that satisfy the spending function
}
