% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.increments.R
\name{get.increments}
\alias{get.increments}
\title{Function to generate increments}
\usage{
get.increments(x)
}
\arguments{
\item{x}{An array is passed from get.boundaries}
}
\description{
This function is not generally used, but is used in the creation of increments
}
