% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.integral.R
\name{get.integral}
\alias{get.integral}
\title{A function to get the integral from an integrand, upper limit, lower limit, and integrand arguments}
\usage{
get.integral(integrand, lower, upper, integrand.arguments)
}
\arguments{
\item{integrand}{integrand given from max distribution and max density functions}

\item{lower}{lower limit in the integral}

\item{upper}{upper limit in the integral}

\item{integrand.arguments}{arguments used in the integral function}
}
\description{
This function isn't generally used, but is used in getting density and distribution
}
