% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max.density.R
\name{get.max.density}
\alias{get.max.density}
\title{Function to get max density}
\usage{
get.max.density(k, incv, x)
}
\arguments{
\item{k}{The number of different treatment groups}

\item{incv}{The number of steps in the integration}

\item{x}{The mean in the dnorm density}
}
\description{
This function performs integration to return density function for maximum of k correlated normals
}
