% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max.distribution.R
\name{get.max.distribution}
\alias{get.max.distribution}
\title{Get Max Distribution Function}
\usage{
get.max.distribution(k, incv, x)
}
\arguments{
\item{k}{The number of treaments being tested in tehe trial}

\item{incv}{The number of integration points}

\item{x}{The points in the integration}
}
\description{
This function isn't generally used. Used to perform integration to return distribution fucntion for K correlated normal distributions
Returns the normal distribution with no integration if k = 1
}
