% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.z.v.Current.R
\name{get.z.v.Current}
\alias{get.z.v.Current}
\title{Get Z and V function for any look}
\usage{
get.z.v.Current(data, n.looks, look, z.v.prev = NULL)
}
\arguments{
\item{data}{Data containing the treatment assignments, short-term outcomes, and long-term outcomes}

\item{n1}{The number of patients with data available for the long-term endpoint at the current analysis}

\item{N2}{The number of patients with data available for the short-term endpoint at the current analysis}

\item{best}{Best is the selected treatment if beyond the first analysis. If at first analysis, then best equals 0.}
}
\description{
This function returns the Z and V for the current look.
}
