% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.z.v.knownsigma.R
\name{get.z.v.knownsigma}
\alias{get.z.v.knownsigma}
\title{get.z.v.knownsigma function}
\usage{
get.z.v.knownsigma(data, n1, N1, N, sigma, rho)
}
\arguments{
\item{data}{data set with primary and secondary endpoint, and treatment}

\item{n1}{Number of patients with primary endpoint available at first analysis}

\item{N1}{Number of patients with secondary endpoint available at first analysis}

\item{N}{The total number of patients in the trial}

\item{sigma}{is the known sigma for the population}

\item{rho}{is the known correlation between endpoints.}
}
\description{
gets z and v for all treatments at first look and for selected treatment at second look using known values for sigma and rho at first look.
Selection is of best treatment from first look
}
