% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.z.v.R
\name{get.z.v}
\alias{get.z.v}
\title{get.z.v function}
\usage{
get.z.v(data, n1, N1, N)
}
\arguments{
\item{data}{data set with primary and secondary endpoint, and treatment}

\item{n1}{Number of patients with primary endpoint available at first analysis}

\item{N1}{Number of patients with secondary endpoint available at first analysis}

\item{N}{The total number of patients in the trial}
}
\description{
gets z and v for all treatments at first look and for selected treatment at second look
}
\examples{
()
get.z.v()
}
