% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.density.integrand.R
\name{max.density.integrand}
\alias{max.density.integrand}
\title{Function to calculate the max desnity integrand}
\usage{
\method{max}{density.integrand}(y, integrand.arguments)
}
\arguments{
\item{y}{The value of the maxmium test statistic in this density}

\item{integrand.arguments}{The integrand arguments passed to the function including k, incv, and x}
}
\description{
This function is not generally used. It returns the integrand used in the max.density calculation
}
