% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.distribution.integrand.R
\name{max.distribution.integrand}
\alias{max.distribution.integrand}
\title{Function for the max distribution integrand}
\usage{
\method{max}{distribution.integrand}(y, integrand.arguments)
}
\arguments{
\item{y}{value of maximum test statistic}

\item{integrand.arguments}{The integrand arguments passed to the function that contain K, incv, and x}
}
\description{
This function is not generally used. This function is used by seamlessTrials to calculate the max distribution integrand
}
