% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARfunctions.R
\name{nocar}
\alias{nocar}
\title{No CAR Function}
\usage{
nocar(covValues, best = 0, tr = NULL, n.trt)
}
\arguments{
\item{covValues}{covValues is a matrix of covariate values, ranging from 0 to total number of factor levels}

\item{best}{best is the best treatment after the first run}

\item{tr}{Tr is vector of treatment assignments from the previous run if applicable}

\item{n.trt}{n.trt is the number of treatments used in the analysis. Only applicable before the first analysis}
}
\description{
This function is to randomally allocation treatments
}
