% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.boundary.values.R
\name{save.boundary.values}
\alias{save.boundary.values}
\title{Function to create the boudnaries that satisfy the spending function}
\usage{
save.boundary.values()
}
\description{
calculate and store boundary values
if no stopping, single critical value/sqrt(v2) just depend on v1/v2
}
