% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatetrials.R
\name{simCarOrig}
\alias{simCarOrig}
\title{Simualte for no CAR}
\usage{
simCarOrig(n1 = 20, N1 = 100, N = 200, n.trt = 3, mean.s = NULL,
  mean.t = NULL, p1 = 0.5, p2 = 0.5, sigma0 = 1, sigma = 1,
  rho = 0.5, nsim = 10000, tau1 = 1, tau2 = 1, design = "Pocock",
  save.boundary)
}
\arguments{
\item{N1}{Number of patients with secondary endpoint available at first analysis}

\item{N}{The total number of patients in the trial}

\item{n.trt}{The number of treatments in the trial}

\item{mean.s}{The mean for short term endpoint sample groups}

\item{mean.t}{Mean for the long term endpoint sample groups}

\item{p1}{The covariate value for the first covariate}

\item{p2}{THe covariate value for the second covariate}

\item{sigma0}{sigma0 in the bivariate normal distribution}

\item{sigma}{is the known sigma for the population}

\item{rho}{is the known correlation between endpoints.}

\item{nsim}{The number of simulation runs. The default is 10000}

\item{tau1}{The chosen covariate adaptive randomization procedure. Default is Pocock's design}

\item{tau2}{The chosen covariate adaptive randomization procedure. Default is Pocock's design}
}
\description{
This function doesn't use an actual second point, just the projected based on the initial point
}
