% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedata.car.R
\name{simulatedata.nocar}
\alias{simulatedata.nocar}
\title{Simulate Data For CAR Function}
\usage{
simulatedata.nocar(mean.s, mean.t, sigma0, sigma, rho, tau1, tau2, treat,
  covValues, data = NULL, inspection)
}
\arguments{
\item{mean.s}{Mean of the short term response}

\item{mean.t}{Mean long-term response}

\item{sigma0}{sigma0 in the bivariate normal distribution}

\item{sigma}{sigma in bivariate normal distribution}

\item{rho}{rho in bivariate normal distribution}

\item{tau1}{Tau1 is the weight on the first covariate}

\item{tau2}{Tau2 is the weight on the second covariate}

\item{treat}{List containing the treatment assignments and the base matrix}

\item{covValues}{covValues is the dummy variable matrix from the covariate values}

\item{data}{Is the data from the previous analysis if applicable. The default is NULL}
}
\description{
Function to simulate data with treatment assignments from Covariate Adaptive Randomization Scheme
}
