% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{simulateData}
\alias{simulateData}
\title{Simulate Data For CAR Function}
\usage{
simulateData(mean.s, mean.t, N = 500, sigma0, sigma, p1, p2, rho,
  design = "Pocock", tau1, tau2)
}
\arguments{
\item{mean.s}{Mean of the short term response}

\item{mean.t}{Mean long-term response}

\item{N}{Total Sample Size}

\item{sigma0}{sigma0 in the bivariate normal distribution}

\item{sigma}{sigma in bivariate normal distribution}

\item{p1}{Probability of success in first covariate}

\item{p2}{Probability of success in second covariate}

\item{rho}{rho in bivariate normal distribution}

\item{design}{Covariate Adapative Randomization scheme. Default is Pocock}

\item{tau1}{Tau1 is the weight on the first covariate}

\item{tau2}{Tau2 is the weight on the second covariate}
}
\description{
Function to simulate data with treatment assignments from Covariate Adaptive Randomization Scheme
}
