% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatetrials.R
\name{simulateest}
\alias{simulateest}
\title{Testing other simulate trials function}
\usage{
simulateest(n1 = 20, N1 = 100, N = 200, n.trt = 3, mean.s = NULL,
  mean.t = NULL, p1, p2, sigma0, sigma, rho, nsim, design = "Pocock",
  tau1, tau2, save.boundary)
}
\arguments{
\item{N1}{Number of patients with secondary endpoint available at first analysis}

\item{N}{The total number of patients in the trial}

\item{n.trt}{The number of treatments in the trial}

\item{mean.s}{The mean for short term endpoint sample groups}

\item{mean.t}{Mean for the long term endpoint sample groups}

\item{p1}{The covariate value for the first covariate}

\item{p2}{THe covariate value for the second covariate}

\item{sigma0}{sigma0 in the bivariate normal distribution}

\item{sigma}{is the known sigma for the population}

\item{rho}{is the known correlation between endpoints.}

\item{nsim}{The number of simulation runs. The default is 1000}

\item{design}{The chosen covariate adaptive randomization procedure. Default is Pocock's design}

\item{tau1}{The chosen covariate adaptive randomization procedure. Default is Pocock's design}

\item{tau2}{The chosen covariate adaptive randomization procedure. Default is Pocock's design}
}
\description{
This function doesn't use an actual second point, just the projected based on the initial point
simulate.trials <- function(n1=20, N1=200, N=200, n.trt=3, mean.s=rep(0,3), mean.t=rep(0,3), p1, sigma0=1, sigma=1, rho=0.0, nsim=1000, save.boundary, simonly=0)
}
